+++ Slider/Assets/Scripts/Map/Desert/DesertCasino.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: f4a1c40c6649b0f4e859b37b2dc1808f
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
