+++ Slider/Assets/Scripts/Map/Desert/DesertTemple.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 8aaeb6d53923af34685f16db1a62f865
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
