+++ Slider/Assets/Scripts/Map/Desert/desertScrollFragTabHintPrompt.cs
@@ -2,7 +2,7 @@ using System.Collections;
 using System.Collections.Generic;
 using UnityEngine;
 
-public class desertScrollFragTabHintPrompt : MonoBehaviour
+public class DesertScrollFragTabHintPrompt : MonoBehaviour
 {
     private bool promptedPreviously = false;
