+++ Slider/Assets/Scripts/Sliders/Grid/AreaGrids/DesertGrid.cs
@@ -15,9 +15,9 @@ public class DesertGrid : SGrid
     public DiceGizmo dice2;
     public SpriteRenderer[] casinoCeilingSprites;
     public List<Animator> casinoSigns;
-    [SerializeField] private ArtifactHousingButtonsManager artifactHousingButtonsManager;
-    [SerializeField] private GameObject templeTrapBlockingRoom;
-    [SerializeField] private GameObject templeTrapBlockingRoomCollider;
+//    [SerializeField] private ArtifactHousingButtonsManager artifactHousingButtonsManager;
+//     [SerializeField] private GameObject templeTrapBlockingRoom;
+//     [SerializeField] private GameObject templeTrapBlockingRoomCollider;
     // [SerializeField] private Collider2D portalCollider; //Desert Portal
     // [SerializeField] private MagiLaser portalLaser;
 
@@ -63,11 +63,11 @@ public class DesertGrid : SGrid
             campfire.SetBool("isDying", false);
         }
 
-        if (SaveSystem.Current.GetBool("desertTempleActivatedTrap") &&
-            !SaveSystem.Current.GetBool("desertTempleTrapCleared"))
-        {
-            ArtifactTabManager.AfterScrollRearrage += OnScrollRearrage;
-        }
+        // if (SaveSystem.Current.GetBool("desertTempleActivatedTrap") &&
+        //     !SaveSystem.Current.GetBool("desertTempleTrapCleared"))
+        // {
+        //     ArtifactTabManager.AfterScrollRearrage += OnScrollRearrage;
+        // }
 
         // portalCollider.enabled = portalEnabled;
         // portalLaser.isEnabled = portalLaserEnabled;
@@ -80,10 +80,10 @@ public class DesertGrid : SGrid
             UIArtifact.OnButtonInteract -= SGrid.UpdateButtonCompletions;
         }
 
-        if (SaveSystem.Current.GetBool("desertTempleActivatedTrap"))
-        {
-            ArtifactTabManager.AfterScrollRearrage -= OnScrollRearrage;
-        }
+        // if (SaveSystem.Current.GetBool("desertTempleActivatedTrap"))
+        // {
+        //     ArtifactTabManager.AfterScrollRearrage -= OnScrollRearrage;
+        // }
     }
 
     private void Update() 
@@ -190,13 +190,11 @@ public class DesertGrid : SGrid
 
         campfireIsLit = profile.GetBool("desertCamp");
         checkCompletion = profile.GetBool("desertCheckCompletion");
-        // portalEnabled = profile.GetBool("magiTechDesertPortal");
-        // portalLaserEnabled = profile.GetBool("magiTechDesertLaser");
 
-        if (SaveSystem.Current.GetBool("desertIsInTemple"))
-        {
-            SetIsInTemple(true);
-        }
+        // if (SaveSystem.Current.GetBool("desertIsInTemple"))
+        // {
+        //     SetIsInTemple(true);
+        // }
     }
 
     // === Desert puzzle specific ===
@@ -216,17 +214,17 @@ public class DesertGrid : SGrid
     }
     #endregion
 
-    #region Jackal
-    //Puzzle 3: Jackal Bone
-    public void CheckJackalNearOasis(Condition c) // no longer used
-    {
-       c.SetSpec(CheckGrid.contains(GetGridString(), "24") || CheckGrid.contains(GetGridString(), "2...4"));
-    }
-    public void CheckDinoNearArch(Condition c)
-    {
-        c.SetSpec(CheckGrid.contains(GetGridString(), "14") || CheckGrid.contains(GetGridString(), "1...4"));
-    }
-    #endregion
+    // #region Jackal
+    // //Puzzle 3: Jackal Bone
+    // public void CheckJackalNearOasis(Condition c) // no longer used
+    // {
+    //    c.SetSpec(CheckGrid.contains(GetGridString(), "24") || CheckGrid.contains(GetGridString(), "2...4"));
+    // }
+    // public void CheckDinoNearArch(Condition c)
+    // {
+    //     c.SetSpec(CheckGrid.contains(GetGridString(), "14") || CheckGrid.contains(GetGridString(), "1...4"));
+    // }
+    // #endregion
 
     #region DicePuzzle
     //Puzzle 4: Dice. Should not start checking until after both tiles have been activated
@@ -256,62 +254,11 @@ public class DesertGrid : SGrid
 
     #endregion
 
-    #region VIPWater
-    //Puzzle 5: Cactus Juice
-    public void HasBottle(Condition c)
-    {
-        c.SetSpec(!SaveSystem.Current.GetBool("desertVIP") &&
-            Player.GetPlayerAction().pickedItem != null && 
-            Player.GetPlayerAction().pickedItem.itemName.Equals("Bottle"));
-    }
-    public void IsCactusJuice(Condition c)
-    {
-        Item item = Player.GetPlayerAction().pickedItem;
-        if (item != null && item.itemName.Equals("Bottle"))
-        {
-            Bottle cast = (Bottle)item;
-            c.SetSpec(cast.state == bottleState.cactus);
-        }
-        else
-        {
-            c.SetSpec(false);
-        }
-    }
-    public void IsDirtyWater(Condition c)
-    {
-        Item item = Player.GetPlayerAction().pickedItem;
-        if (item != null && item.itemName.Equals("Bottle"))
-        {
-            Bottle cast = (Bottle)item;
-            c.SetSpec(cast.state == bottleState.dirty);
-        }
-        else
-        {
-            c.SetSpec(false);
-        }
-    }
-    public void IsCleanWater(Condition c)
-    {
-        Item item = Player.GetPlayerAction().pickedItem;
-        if (item != null && item.itemName.Equals("Bottle"))
-        {
-            Bottle cast = (Bottle)item;
-            c.SetSpec(cast.state == bottleState.clean || SaveSystem.Current.GetBool("desertVIP"));            
-        }
-        else
-        {
-            c.SetSpec(SaveSystem.Current.GetBool("desertVIP"));
-        }
-    }
-    #endregion
-
     #region Gazelle
-    //Puzzle 6: Shady Gazelle
     public void CheckGazelleNearOasis(Condition c)
     {
         c.SetSpec(CheckGrid.contains(GetGridString(), "26") || CheckGrid.contains(GetGridString(), "6...2"));
     }
-
     #endregion
 
     #region Party
@@ -351,10 +298,6 @@ public class DesertGrid : SGrid
 
     private IEnumerator ShuffleBuildUp()
     {
-        //AudioManager.Play("Puzzle Complete");
-
-        //yield return new WaitForSeconds(0.5f);
-
         CameraShake.Shake(0.25f, 0.25f);
         AudioManager.Play("Slide Rumble");
 
@@ -386,9 +329,9 @@ public class DesertGrid : SGrid
         shuffleBuildUpCoroutine = null;
     }
 
-    public void GiveScrollAchievement() {
-        AchievementManager.SetAchievementStat("collectedScroll", 1);
-    }
+    // public void GiveScrollAchievement() {
+    //     AchievementManager.SetAchievementStat("collectedScroll", 1);
+    // }
 
     private void DoShuffle()
     {
@@ -446,78 +389,78 @@ public class DesertGrid : SGrid
 
     #endregion
 
-    #region Scroll
+    // #region Scroll
 
-    public void SetIsInTemple(bool isInTemple)
-    {
-        SaveSystem.Current.SetBool("desertIsInTemple", isInTemple);
-        if (isInTemple)
-        {
-            SaveSystem.Current.SetBool("desertEnteredTemple", true);
-        }
+    // public void SetIsInTemple(bool isInTemple)
+    // {
+    //     SaveSystem.Current.SetBool("desertIsInTemple", isInTemple);
+    //     if (isInTemple)
+    //     {
+    //         SaveSystem.Current.SetBool("desertEnteredTemple", true);
+    //     }
 
-        artifactHousingButtonsManager.SetSpritesToHousing(isInTemple);
-        Player._instance.SetTracker(!isInTemple);
-        Player._instance.SetDontUpdateSTileUnderneath(isInTemple);
-    }
+    //     artifactHousingButtonsManager.SetSpritesToHousing(isInTemple);
+    //     Player._instance.SetTracker(!isInTemple);
+    //     Player._instance.SetDontUpdateSTileUnderneath(isInTemple);
+    // }
 
-    public void ActivateTrap()
-    {
-        SaveSystem.Current.SetBool("desertTempleTrapActivated", true);
-        if (shuffleBuildUpCoroutine == null)
-        {
-            shuffleBuildUpCoroutine = StartCoroutine(ActivateTrapBuildUp());
-        }
-    }
+    // public void ActivateTrap()
+    // {
+    //     SaveSystem.Current.SetBool("desertTempleTrapActivated", true);
+    //     if (shuffleBuildUpCoroutine == null)
+    //     {
+    //         shuffleBuildUpCoroutine = StartCoroutine(ActivateTrapBuildUp());
+    //     }
+    // }
 
-    private IEnumerator ActivateTrapBuildUp()
-    {
-        templeTrapBlockingRoomCollider.SetActive(true);
+    // private IEnumerator ActivateTrapBuildUp()
+    // {
+    //     templeTrapBlockingRoomCollider.SetActive(true);
 
-        CameraShake.Shake(0.25f, 0.25f);
-        AudioManager.Play("Slide Rumble");
+    //     CameraShake.Shake(0.25f, 0.25f);
+    //     AudioManager.Play("Slide Rumble");
 
-        yield return new WaitForSeconds(1f);
+    //     yield return new WaitForSeconds(1f);
 
-        CameraShake.Shake(0.25f, 0.25f);
-        AudioManager.Play("Slide Rumble");
+    //     CameraShake.Shake(0.25f, 0.25f);
+    //     AudioManager.Play("Slide Rumble");
 
-        yield return new WaitForSeconds(1f);
+    //     yield return new WaitForSeconds(1f);
 
-        CameraShake.Shake(0.75f, 0.5f);
-        AudioManager.Play("Slide Rumble");
+    //     CameraShake.Shake(0.75f, 0.5f);
+    //     AudioManager.Play("Slide Rumble");
 
-        yield return new WaitForSeconds(1f);
+    //     yield return new WaitForSeconds(1f);
 
-        CameraShake.Shake(1.5f, 2.5f);
-        AudioManager.PlayWithVolume("Slide Explosion", 0.2f);
-        AudioManager.Play("TFT Bell");
+    //     CameraShake.Shake(1.5f, 2.5f);
+    //     AudioManager.PlayWithVolume("Slide Explosion", 0.2f);
+    //     AudioManager.Play("TFT Bell");
 
-        yield return new WaitForSeconds(0.25f);
+    //     yield return new WaitForSeconds(0.25f);
 
-        UIEffects.FlashWhite();
-        SGrid.Current.SetGrid(SGrid.GridStringToSetGridFormat("815493672"));
-        templeTrapBlockingRoom.SetActive(true);
-        SaveSystem.Current.SetBool("desertTempleActivatedTrap", true);
+    //     UIEffects.FlashWhite();
+    //     SGrid.Current.SetGrid(SGrid.GridStringToSetGridFormat("815493672"));
+    //     templeTrapBlockingRoom.SetActive(true);
+    //     SaveSystem.Current.SetBool("desertTempleActivatedTrap", true);
 
-        ArtifactTabManager.AfterScrollRearrage += OnScrollRearrage;
+    //     ArtifactTabManager.AfterScrollRearrage += OnScrollRearrage;
 
-        yield return new WaitForSeconds(0.75f);
+    //     yield return new WaitForSeconds(0.75f);
 
-        CameraShake.Shake(2, 0.9f);
-        shuffleBuildUpCoroutine = null;
-    }
+    //     CameraShake.Shake(2, 0.9f);
+    //     shuffleBuildUpCoroutine = null;
+    // }
 
-    private void OnScrollRearrage(object sender, System.EventArgs e)
-    {
-        templeTrapBlockingRoom.SetActive(false);
-        templeTrapBlockingRoomCollider.SetActive(false);
-        SaveSystem.Current.SetBool("desertTempleTrapCleared", true);
-        ArtifactTabManager.AfterScrollRearrage -= OnScrollRearrage;
+    // private void OnScrollRearrage(object sender, System.EventArgs e)
+    // {
+    //     templeTrapBlockingRoom.SetActive(false);
+    //     templeTrapBlockingRoomCollider.SetActive(false);
+    //     SaveSystem.Current.SetBool("desertTempleTrapCleared", true);
+    //     ArtifactTabManager.AfterScrollRearrage -= OnScrollRearrage;
 
-        CheckFinalPlacements(UIArtifact.GetGridString());
-        AchievementManager.SetAchievementStat("completedDesert", 1);
-    }
+    //     CheckFinalPlacements(UIArtifact.GetGridString());
+    //     AchievementManager.SetAchievementStat("completedDesert", 1);
+    // }
 
-    #endregion
+    // #endregion
 }
